SearchView refactored
