+++ src/domain/challenge/challenge/containers/ChallengeExplorerContainer.tsx
@@ -10,7 +10,7 @@ import { useUserContext } from '../../../community/contributor/user';
 import { ValueType } from '../../../../common/components/core/card-filter/filterFn';
 import { getVisualBannerNarrow } from '../../../common/visual/utils/visuals.utils';
 import { SearchResultChallengeFragment } from '../../../../core/apollo/generated/graphql-schema';
-import { SearchResultT } from '../../../platform/pages/Search/SearchPage';
+import { SearchResultT } from '../../../platform/search/SearchView';
 
 export type SimpleChallenge = {
   id: string;
