+++ src/domain/platform/pages/Search/SearchResultCardChooser.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Skeleton from '@mui/material/Skeleton';
 import { useHydrateCard } from './hooks/useHydratedCard';
-import { SearchResultMetaType } from './SearchPage';
+import { SearchResultMetaType } from '../../search/SearchView';
 import { SearchResultType } from '../../../../core/apollo/generated/graphql-schema';
 
 const SearchResultCardChooser = ({
