--- src/domain/platform/pages/Search/index.ts
@@ -1,1 +1,0 @@
-export * from './SearchPage';
